+++ php/src/Indexer.php
@@ -129,8 +129,6 @@ class Indexer
      * Indexes the specified project.
      *
      * @param string $directory
-     *
-     * @return array A list of errors accumulated during indexing (indexing continues when these happen).
      */
     public function indexDirectory($directory)
     {
@@ -146,23 +144,17 @@ class Indexer
 
         $this->sendProgress(0, $totalItems);
 
-        $errors = [];
-
         foreach ($files as $i => $filePath) {
             echo $this->logMessage('  - Indexing ' . $filePath);
 
             try {
                 $this->indexFile($filePath);
             } catch (Indexer\IndexingFailedException $e) {
-                $errors = array_merge($errors, $e->getErrors());
-
                 $this->logMessage('    - ERROR: Indexing failed due to parsing errors!');
             }
 
             $this->sendProgress($i+1, $totalItems);
         }
-
-        return $errors;
     }
 
     /**
@@ -194,16 +186,7 @@ class Indexer
             $traverser->addVisitor($useStatementFetchingVisitor);
             $traverser->traverse($nodes);
         } catch (Error $e) {
-            throw new Indexer\IndexingFailedException([
-                [
-                    'file'        => $filePath,
-                    'startLine'   => $e->getStartLine() >= 0 ? $e->getStartLine() : null,
-                    'endLine'     => $e->getEndLine() >= 0 ? $e->getEndLine() : null,
-                    'startColumn' => $e->hasColumnInfo() ? $e->getStartColumn($code) : null,
-                    'endColumn'   => $e->hasColumnInfo() ? $e->getEndColumn($code) : null,
-                    'message'     => $e->getMessage()
-                ]
-            ]);
+            throw new Indexer\IndexingFailedException();
         }
 
         $this->storage->beginTransaction();