+++ php/src/Indexer/IndexingFailedException.php
@@ -9,28 +9,5 @@ use RuntimeException;
  */
 class IndexingFailedException extends RuntimeException
 {
-    /**
-     * @var array
-     */
-    protected $errors;
 
-    /**
-     * Constructor.
-     *
-     * @param array $errors.
-     */
-    public function __construct(array $errors = [])
-    {
-        $this->errors = $errors;
-    }
-
-    /**
-     * Retrieves the currently set errors.
-     *
-     * @return array
-     */
-    public function getErrors()
-    {
-        return $this->errors;
-    }
 }