+++ php/tests/Application/Command/SemanticLintTest/SyntaxError.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace A;
+
+$test = 5;
+$syntaxError = ;
+$anotherTest = 2;
+$anotherSyntaxError = ;
+$lastTest = 3;